TS BUG: need declaration: false to build correctly
If we set declaration: true, then typescript (specifically rpt2,
Rollup's TS plugin) fails with a cryptic path error. This is
filed under https://github.com/Microsoft/TypeScript/issues/25047.

Curiously, this does not happen on master, and I have vaguely
traced it back to unionize and the fact that it does fancy type
exports, which we partially re-export here. We would have to
manually emit a declaration file, which is not hard tbh, but a bit
annoying. Even funnier, webpack has no issue building in storybook,
but that might be because it does not emit declarations anyway.

Steps:
- Get declaration file from master, add to repository, and copy to
dist/ upon build;
- See if massaging our exports fixes it;
- Wait for TS :/